requirements.txt
ttkbootstrap>=1.10.0
matplotlib>=3.5.0
numpy>=1.21.0
scipy>=1.7.0
spidev>=3.5
RPi.GPIO>=0.7.0
